fix(switch-off): problem with logout
Trying close BSWPM on QTILE was ended failed. I added conditions and that fixed problem

+++ pkgs/utils/rofiSwitchOff/default.nix
@@ -17,15 +17,21 @@
       ${systemd}/bin/poweroff
       ;;
     "$LOGOUT")
-      QTILE=$(pgrep qtile)
-      BSPWM=$(pgrep bspwm)
-      kill -9 "$QTILE"
-      kill -9 "$BSPWM"
+      QTILE=$(${toybox}/bin/pgrep qtile)
+      if [ -n "$QTILE" ]; then
+        echo "Kill qtile $QTILE"
+        kill -9 "$QTILE"
+      fi 
+      BSPWM=$(${toybox}/bin/pgrep bspwm)
+      if [ -n "$BSPWM" ]; then
+        kill -9 "$BSPWM"
+      fi
       ;;
     "$REBOOT")
       ${systemd}/bin/reboot
       ;;
     "*")
+      echo "invalid option"
       exit 1
       ;;
     esac
